requirements.txt
numpy>=1.22.1
opencv-python>=4.4.0
Pillow>=7.1.2
onnxruntime>=1.11.1